requirements.txt
flask>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
werkzeug>=3.0.0
nest-asyncio>=1.5.0
gunicorn>=21.0.0